Clean up common lib code
